apps/books/views.py
from rest_framework import generics
from rest_framework.permissions import IsAuthenticated
from .models import Book
from .serializers import BookListSerializer

class BookListView(generics.ListCreateAPIView):
    permission_classes = [IsAuthenticated]
    queryset = Book.objects.all().order_by('-created_at')
    serializer_class = BookListSerializer
